+++ src/app/(root)/haku-table.tsx
@@ -1,12 +1,10 @@
 'use client';
-import ListTable, {
-  ListTableColumn,
-  makeCountColumn,
-} from '../components/table/list-table';
+import ListTable, { makeCountColumn } from '../components/table/list-table';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { HakuListItem } from '@/app/hooks/useHakuSearch';
 import { Link } from '@mui/material';
 import { useMemo } from 'react';
+import { ListTableColumn } from '../components/table/table-types';
 
 export const HakuTable = ({
   haut,
