+++ src/app/components/table/list-table.tsx
@@ -2,8 +2,6 @@
 
 import {
   Box,
-  Button,
-  Checkbox,
   Table,
   TableBody,
   TableCell,
@@ -11,27 +9,16 @@ import {
   TableRow,
   styled,
 } from '@mui/material';
-import { ExpandLess, ExpandMore, UnfoldMore } from '@mui/icons-material';
-import { getSortParts } from './table-utils';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { TFunction } from 'i18next';
 import { ExternalLink } from '../external-link';
 import { useMemo } from 'react';
 import { OphPagination } from './oph-pagination';
-import { map, pipe } from 'remeda';
 import { ophColors } from '@opetushallitus/oph-design-system';
-
-type KeysMatching<O, T> = {
-  [K in keyof O]: O[K] extends T ? K : never;
-}[keyof O & string];
-
-export type ListTableColumn<P> = {
-  title?: string;
-  key: keyof P;
-  render: (props: P) => React.ReactNode;
-  style?: React.CSSProperties;
-  sortable?: boolean;
-};
+import { TableHeaderCell } from './table-header-cell';
+import { EMPTY_ARRAY, EMPTY_STRING_SET } from '@/app/lib/common';
+import { TableHeaderCheckbox, TableRowCheckbox } from './table-checkboxes';
+import { KeysMatching, ListTableColumn, Row } from './table-types';
 
 export const makeGenericColumn = <T extends Record<string, unknown>>({
   title,
@@ -163,14 +150,6 @@ const StyledCell = styled(TableCell)({
   },
 });
 
-const StyledHeaderCell = styled(TableCell)({
-  padding: '0.2rem 0.1rem 0.2rem 0.4rem',
-  textAlign: 'left',
-  'button:focus': {
-    color: ophColors.blue2,
-  },
-});
-
 const StyledTableBody = styled(TableBody)({
   '& .MuiTableRow-root': {
     '&:nth-of-type(even)': {
@@ -210,69 +189,6 @@ interface ListTableProps<T extends Row>
   onSelectionChange?: (selection: Set<string>) => void;
 }
 
-const SortIcon = ({
-  sortValue,
-  colId = '',
-}: {
-  sortValue?: string;
-  colId?: string;
-}) => {
-  switch (sortValue) {
-    case `${colId}:asc`:
-      return <ExpandLess />;
-    case `${colId}:desc`:
-      return <ExpandMore />;
-    default:
-      return <UnfoldMore />;
-  }
-};
-
-const HeaderCell = ({
-  colId,
-  title,
-  style,
-  sort,
-  setSort,
-  sortable,
-}: {
-  colId?: string;
-  title?: React.ReactNode;
-  style?: React.CSSProperties;
-  sort?: string;
-  setSort?: (sort: string) => void;
-  sortable?: boolean;
-}) => {
-  const { direction } = getSortParts(sort, colId);
-
-  return (
-    <StyledHeaderCell sx={style} sortDirection={direction}>
-      {setSort && sortable ? (
-        <Button
-          sx={{
-            color: ophColors.black,
-          }}
-          onClick={() => {
-            let newSortValue = '';
-            if (sort === `${colId}:asc`) {
-              newSortValue = `${colId}:desc`;
-            } else if (sort === `${colId}:desc`) {
-              newSortValue = '';
-            } else {
-              newSortValue = `${colId}:asc`;
-            }
-            setSort(newSortValue);
-          }}
-          endIcon={<SortIcon sortValue={sort} colId={colId} />}
-        >
-          {title}
-        </Button>
-      ) : (
-        <span style={{ fontWeight: 600 }}>{title}</span>
-      )}
-    </StyledHeaderCell>
-  );
-};
-
 const TableWrapper = styled(Box)(({ theme }) => ({
   position: 'relative',
   display: 'flex',
@@ -282,12 +198,6 @@ const TableWrapper = styled(Box)(({ theme }) => ({
   overflowX: 'auto',
   rowGap: theme.spacing(1),
 }));
-
-const ListCheckbox = styled(Checkbox)({
-  padding: 0,
-});
-
-export type Row<K extends string = string> = Record<K, unknown>;
 
 const TablePagination = ({
   label,
@@ -316,9 +226,6 @@ const TablePagination = ({
   );
 };
 
-const EMPTY_SET: Set<string> = new Set();
-const EMPTY_ARRAY: Array<unknown> = [];
-
 export const ListTable = <T extends Row>({
   columns = EMPTY_ARRAY as Array<ListTableColumn<T>>,
   rows = EMPTY_ARRAY as Array<T>,
@@ -328,7 +235,7 @@ export const ListTable = <T extends Row>({
   translateHeader = true,
   pagination,
   checkboxSelection,
-  selection = EMPTY_SET,
+  selection = EMPTY_STRING_SET,
   onSelectionChange,
   getRowCheckboxLabel,
   ...props
@@ -349,33 +256,14 @@ export const ListTable = <T extends Row>({
         <TableHead>
           <TableRow sx={{ borderBottom: `2px solid ${ophColors.grey200}` }}>
             {checkboxSelection && (
-              <HeaderCell
+              <TableHeaderCell
                 key="select-all"
                 title={
-                  <ListCheckbox
-                    checked={selection.size === rows.length}
-                    indeterminate={
-                      selection.size > 0 &&
-                      selection.size !== rows.length &&
-                      selection.size !== 0
-                    }
-                    inputProps={{ 'aria-label': t('yleinen.valitse-kaikki') }}
-                    onChange={(
-                      event: React.ChangeEvent<HTMLInputElement>,
-                      checked: boolean,
-                    ) => {
-                      if (checked) {
-                        onSelectionChange?.(
-                          pipe(
-                            rows,
-                            map((item) => item?.[rowKeyProp] as string),
-                            (array) => new Set(array),
-                          ),
-                        );
-                      } else {
-                        onSelectionChange?.(EMPTY_SET);
-                      }
-                    }}
+                  <TableHeaderCheckbox
+                    selection={selection}
+                    onSelectionChange={onSelectionChange}
+                    rows={rows}
+                    rowKeyProp={rowKeyProp}
                   />
                 }
               />
@@ -383,7 +271,7 @@ export const ListTable = <T extends Row>({
             {columns.map((columnProps) => {
               const { key, title, style, sortable } = columnProps;
               return (
-                <HeaderCell
+                <TableHeaderCell
                   key={key.toString()}
                   colId={key.toString()}
                   title={translateHeader ? t(title ?? '') : title}
@@ -403,25 +291,12 @@ export const ListTable = <T extends Row>({
               <TableRow key={rowId}>
                 {checkboxSelection && (
                   <StyledCell>
-                    <ListCheckbox
-                      checked={selection.has(rowId)}
-                      inputProps={{
-                        'aria-label': getRowCheckboxLabel?.(rowProps),
-                      }}
-                      value={rowId}
-                      onChange={(
-                        event: React.ChangeEvent<HTMLInputElement>,
-                        checked: boolean,
-                      ) => {
-                        const newSelection = new Set(selection);
-                        if (checked) {
-                          newSelection.add(rowId);
-                          onSelectionChange?.(newSelection);
-                        } else {
-                          newSelection.delete(rowId);
-                          onSelectionChange?.(newSelection);
-                        }
-                      }}
+                    <TableRowCheckbox
+                      selection={selection}
+                      onSelectionChange={onSelectionChange}
+                      rowId={rowId}
+                      rowProps={rowProps}
+                      getRowCheckboxLabel={getRowCheckboxLabel}
                     />
                   </StyledCell>
                 )}
