+++ src/app/lib/common.ts
@@ -34,6 +34,7 @@ export const isEmpty = (value: unknown) => {
 
 export const EMPTY_OBJECT = Object.freeze({});
 export const EMPTY_ARRAY = Object.freeze([]) as Array<never>;
+export const EMPTY_STRING_SET = Object.freeze(new Set<string>());
 
 export function downloadBlob(fileName: string, data: Blob) {
   const link = document.createElement('a');
